@@ -4,13 +4,12 @@ version: 0.1.1
 homepage: https://github.com/dev-juyoung/Fitness
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.20.0 <2.0.0"
+  sdk: '>=2.12.0 <3.0.0'
+  flutter: ">=1.20.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  meta: ^1.2.4
 
 dev_dependencies:
   flutter_test:
